split .hpp between proper .hpp and .cpp
yeah, i run into the multiple definition error, didn't want to inline everythings. Thus i moved everythings.
