+++ src/main.cpp
@@ -1,3 +1,4 @@
+#pragma once
 #include <iostream>
 #include <GLFW/glfw3.h>
 #include "maths/linear.hpp"
