+++ src/maths/matrix.hpp
@@ -71,159 +71,47 @@ namespace Acamarachi::Maths
 
 	///---		ADDITION     ---///
 	//operator +=
-	void operator+=(mat2& a, mat2 o)
-	{
-		a[0] += o[0];
-		a[1] += o[1];
-	}
-
-	void operator+=(mat3& a, mat3 o)
-	{
-		a[0] += o[0];
-		a[1] += o[1];
-		a[2] += o[2];
-	}
-
-	void operator+=(mat4& a, mat4 o)
-	{
-		a[0] += o[0];
-		a[1] += o[1];
-		a[2] += o[2];
-		a[3] += o[3];
-	}
+	void operator+=(mat2& a, mat2 o);
+	void operator+=(mat3& a, mat3 o);
+	void operator+=(mat4& a, mat4 o);
 
 	//operator +=
-	void operator+=(mat2& a, float o)
-	{
-		a[0] += o;
-		a[1] += o;
-	}
-
-	void operator+=(mat3& a, float o)
-	{
-		a[0] += o;
-		a[1] += o;
-		a[2] += o;
-	}
-
-	void operator+=(mat4& a, float o)
-	{
-		a[0] += o;
-		a[1] += o;
-		a[2] += o;
-		a[3] += o;
-	}
+	void operator+=(mat2& a, float o);
+	void operator+=(mat3& a, float o);
+	void operator+=(mat4& a, float o);
 
 	//matrix + matrix
 	
-	mat2 operator+(mat2 a, mat2 b)
-	{
-		return mat2(a[0] + b[0], a[1] + b[1]);
-	}
-
-	mat3 operator+(mat3 a, mat3 b)
-	{
-		return mat3(a[0] + b[0], a[1] + b[1], a[2] + b[2]);
-	}
-
-	mat4 operator+(mat4 a, mat4 b)
-	{
-		return mat4(a[0] + b[0], a[1] + b[1], a[2] + b[2], a[3] + b[3]);
-	}
+	mat2 operator+(mat2 a, mat2 b);
+	mat3 operator+(mat3 a, mat3 b);
+	mat4 operator+(mat4 a, mat4 b);
 
 	//matrix + float
 
-	mat2 operator+(mat2 a, float b)
-	{
-		return mat2(a[0] + b, a[1] + b);
-	}
-
-	mat3 operator+(mat3 a, float b)
-	{
-		return mat3(a[0] + b, a[1] + b, a[2] + b);
-	}
-
-	mat4 operator+(mat4 a, float b)
-	{
-		return mat4(a[0] + b, a[1] + b, a[2] + b, a[3] + b);
-	}
+	mat2 operator+(mat2 a, float b);
+	mat3 operator+(mat3 a, float b);
+	mat4 operator+(mat4 a, float b);
 
 	///---		SUBSTRACTION     ---///
 	//operator -=
-	void operator-=(mat2& a, mat2 o)
-	{
-		a[0] -= o[0];
-		a[1] -= o[1];
-	}
-
-	void operator-=(mat3& a, mat3 o)
-	{
-		a[0] -= o[0];
-		a[1] -= o[1];
-		a[2] -= o[2];
-	}
-
-	void operator-=(mat4& a, mat4 o)
-	{
-		a[0] -= o[0];
-		a[1] -= o[1];
-		a[2] -= o[2];
-		a[3] -= o[3];
-	}
+	void operator-=(mat2& a, mat2 o);
+	void operator-=(mat3& a, mat3 o);
+	void operator-=(mat4& a, mat4 o);
 
 	//operator -=
-	void operator-=(mat2& a, float o)
-	{
-		a[0] -= o;
-		a[1] -= o;
-	}
-
-	void operator-=(mat3& a, float o)
-	{
-		a[0] -= o;
-		a[1] -= o;
-		a[2] -= o;
-	}
-
-	void operator-=(mat4& a, float o)
-	{
-		a[0] -= o;
-		a[1] -= o;
-		a[2] -= o;
-		a[3] -= o;
-	}
+	void operator-=(mat2& a, float o);
+	void operator-=(mat3& a, float o);
+	void operator-=(mat4& a, float o);
 
 	//matrix - matrix
 
-	mat2 operator-(mat2 a, mat2 b)
-	{
-		return mat2(a[0] - b[0], a[1] - b[1]);
-	}
-
-	mat3 operator-(mat3 a, mat3 b)
-	{
-		return mat3(a[0] - b[0], a[1] - b[1], a[2] - b[2]);
-	}
-
-	mat4 operator-(mat4 a, mat4 b)
-	{
-		return mat4(a[0] - b[0], a[1] - b[1], a[2] - b[2], a[3] - b[3]);
-	}
+	mat2 operator-(mat2 a, mat2 b);
+	mat3 operator-(mat3 a, mat3 b);
+	mat4 operator-(mat4 a, mat4 b);
 
 	//matrix - float
 
-	mat2 operator-(mat2 a, float b)
-	{
-		return mat2(a[0] - b, a[1] - b);
-	}
-
-	mat3 operator-(mat3 a, float b)
-	{
-		return mat3(a[0] - b, a[1] - b, a[2] - b);
-	}
-
-	mat4 operator-(mat4 a, float b)
-	{
-		return mat4(a[0] - b, a[1] - b, a[2] - b, a[3] - b);
-	}
+	mat2 operator-(mat2 a, float b);
+	mat3 operator-(mat3 a, float b);
+	mat4 operator-(mat4 a, float b);
 }
